[DC-1780] Add clustering and mixed insert/update mode (#27)
This PR adds the following functionality:
- Specify number of distinct cluster values to test clustered collections via `NUM_CLUSTERS`
- Specify percentage of documents hitting a specific cluster to simulate hot clustering via `HOT_CLUSTER_PERCENTAGE`
- Add a mode that mixes inserts and updates randomly via `mixed` mode, `UPDATE_PERCENTAGE`, and `MAX_DOCS`

Mixed mode only works well if there's one kubernetes replica doing the work. The max doc id, n, is tracked locally so that any docs with doc id 0 to n will be updates and n+1 will be an insert. The alternative that would work for multiple replicas is to use a uuid on inserts and for updates choose from 0 to n. This means the collection would need to be populated with docs from 0 to n first, and any newly inserted docs with uuids would never be updated.